requirements.txt
openai==1.75.0
scrapingbee==2.0.1
requests==2.32.3
boto3==1.37.37
tiktoken==0.9.0
opencv-python-headless==4.11.0.86
beautifulsoup4==4.13.4
numpy==2.2.4
python-dotenv==1.1.0
typesense==1.0.3